# src/main/java/ru/chousik/is/api/CoordinatesApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.16.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package ru.chousik.is.api;

import ru.chousik.is.dto.request.CoordinatesAddRequest;
import ru.chousik.is.dto.request.CoordinatesUpdateRequest;
import org.springframework.lang.Nullable;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-10-12T18:08:29.283737+03:00[Europe/Moscow]", comments = "Generator version: 7.16.0")
@Validated
@Tag(name = "Coordinates", description = "the Coordinates API")
public interface CoordinatesApi {

    public static final String PATH_API_V1_COORDINATES_BY_IDS_GET = "/api/v1/coordinates/by-ids";
    /**
     * GET /api/v1/coordinates/by-ids : Получить несколько координат
     *
     * @param ids  (required)
     * @return Список координат (status code 200)
     */
    @Operation(
        operationId = "apiV1CoordinatesByIdsGet",
        summary = "Получить несколько координат",
        tags = { "Coordinates" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Список координат", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = ru.chousik.is.dto.response.CoordinatesResponse.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = CoordinatesApi.PATH_API_V1_COORDINATES_BY_IDS_GET,
        produces = { "application/json" }
    )
    
    ResponseEntity<List<ru.chousik.is.dto.response.CoordinatesResponse>> apiV1CoordinatesByIdsGet(
        @NotNull @Parameter(name = "ids", description = "", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "ids", required = true) List<Long> ids
    );


    public static final String PATH_API_V1_COORDINATES_DELETE = "/api/v1/coordinates";
    /**
     * DELETE /api/v1/coordinates : Массово удалить координаты
     *
     * @param ids  (required)
     * @return Координаты удалены (status code 204)
     */
    @Operation(
        operationId = "apiV1CoordinatesDelete",
        summary = "Массово удалить координаты",
        tags = { "Coordinates" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Координаты удалены")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = CoordinatesApi.PATH_API_V1_COORDINATES_DELETE
    )
    
    ResponseEntity<Void> apiV1CoordinatesDelete(
        @NotNull @Parameter(name = "ids", description = "", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "ids", required = true) List<Long> ids
    );


    public static final String PATH_API_V1_COORDINATES_GET = "/api/v1/coordinates";
    /**
     * GET /api/v1/coordinates : Получить список координат
     * Возвращает постраничный список координат с сортировкой.
     *
     * @param page Номер страницы (0-based) (optional)
     * @param size Размер страницы (optional)
     * @param sort Нативный параметр Spring для сортировки (optional)
     * @param sortBy Поле сортировки (optional)
     * @param direction Направление сортировки (optional)
     * @return Список координат (status code 200)
     */
    @Operation(
        operationId = "apiV1CoordinatesGet",
        summary = "Получить список координат",
        description = "Возвращает постраничный список координат с сортировкой.",
        tags = { "Coordinates" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Список координат", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = org.springframework.data.web.PagedModel.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = CoordinatesApi.PATH_API_V1_COORDINATES_GET,
        produces = { "application/json" }
    )
    
    ResponseEntity<org.springframework.data.web.PagedModel> apiV1CoordinatesGet(
        @Min(0) @Parameter(name = "page", description = "Номер страницы (0-based)", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false) @Nullable Integer page,
        @Min(1) @Parameter(name = "size", description = "Размер страницы", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false) @Nullable Integer size,
        @Parameter(name = "sort", description = "Нативный параметр Spring для сортировки", in = ParameterIn.QUERY) @Valid @RequestParam(value = "sort", required = false) @Nullable String sort,
        @Parameter(name = "sortBy", description = "Поле сортировки", in = ParameterIn.QUERY) @Valid @RequestParam(value = "sortBy", required = false) @Nullable String sortBy,
        @Parameter(name = "direction", description = "Направление сортировки", in = ParameterIn.QUERY) @Valid @RequestParam(value = "direction", required = false) @Nullable String direction
    );


    public static final String PATH_API_V1_COORDINATES_ID_DELETE = "/api/v1/coordinates/{id}";
    /**
     * DELETE /api/v1/coordinates/{id} : Удалить координаты
     *
     * @param id  (required)
     * @return Удалённые координаты (status code 200)
     */
    @Operation(
        operationId = "apiV1CoordinatesIdDelete",
        summary = "Удалить координаты",
        tags = { "Coordinates" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Удалённые координаты", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ru.chousik.is.dto.response.CoordinatesResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = CoordinatesApi.PATH_API_V1_COORDINATES_ID_DELETE,
        produces = { "application/json" }
    )
    
    ResponseEntity<ru.chousik.is.dto.response.CoordinatesResponse> apiV1CoordinatesIdDelete(
        @NotNull @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    public static final String PATH_API_V1_COORDINATES_ID_GET = "/api/v1/coordinates/{id}";
    /**
     * GET /api/v1/coordinates/{id} : Получить координаты по идентификатору
     *
     * @param id  (required)
     * @return Найденные координаты (status code 200)
     */
    @Operation(
        operationId = "apiV1CoordinatesIdGet",
        summary = "Получить координаты по идентификатору",
        tags = { "Coordinates" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Найденные координаты", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ru.chousik.is.dto.response.CoordinatesResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = CoordinatesApi.PATH_API_V1_COORDINATES_ID_GET,
        produces = { "application/json" }
    )
    
    ResponseEntity<ru.chousik.is.dto.response.CoordinatesResponse> apiV1CoordinatesIdGet(
        @NotNull @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    public static final String PATH_API_V1_COORDINATES_ID_PATCH = "/api/v1/coordinates/{id}";
    /**
     * PATCH /api/v1/coordinates/{id} : Обновить координаты
     *
     * @param id  (required)
     * @param coordinatesUpdateRequest  (required)
     * @return Обновлённые координаты (status code 200)
     */
    @Operation(
        operationId = "apiV1CoordinatesIdPatch",
        summary = "Обновить координаты",
        tags = { "Coordinates" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Обновлённые координаты", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ru.chousik.is.dto.response.CoordinatesResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PATCH,
        value = CoordinatesApi.PATH_API_V1_COORDINATES_ID_PATCH,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<ru.chousik.is.dto.response.CoordinatesResponse> apiV1CoordinatesIdPatch(
        @NotNull @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "CoordinatesUpdateRequest", description = "", required = true) @Valid @RequestBody CoordinatesUpdateRequest coordinatesUpdateRequest
    );


    public static final String PATH_API_V1_COORDINATES_PATCH = "/api/v1/coordinates";
    /**
     * PATCH /api/v1/coordinates : Массово обновить координаты
     *
     * @param ids  (required)
     * @param coordinatesUpdateRequest  (required)
     * @return Обновлённые координаты (status code 200)
     */
    @Operation(
        operationId = "apiV1CoordinatesPatch",
        summary = "Массово обновить координаты",
        tags = { "Coordinates" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Обновлённые координаты", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = ru.chousik.is.dto.response.CoordinatesResponse.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PATCH,
        value = CoordinatesApi.PATH_API_V1_COORDINATES_PATCH,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<List<ru.chousik.is.dto.response.CoordinatesResponse>> apiV1CoordinatesPatch(
        @NotNull @Parameter(name = "ids", description = "", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "ids", required = true) List<Long> ids,
        @Parameter(name = "CoordinatesUpdateRequest", description = "", required = true) @Valid @RequestBody CoordinatesUpdateRequest coordinatesUpdateRequest
    );


    public static final String PATH_API_V1_COORDINATES_POST = "/api/v1/coordinates";
    /**
     * POST /api/v1/coordinates : Создать координаты
     *
     * @param coordinatesAddRequest  (required)
     * @return Созданные координаты (status code 200)
     */
    @Operation(
        operationId = "apiV1CoordinatesPost",
        summary = "Создать координаты",
        tags = { "Coordinates" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Созданные координаты", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ru.chousik.is.dto.response.CoordinatesResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = CoordinatesApi.PATH_API_V1_COORDINATES_POST,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<ru.chousik.is.dto.response.CoordinatesResponse> apiV1CoordinatesPost(
        @Parameter(name = "CoordinatesAddRequest", description = "", required = true) @Valid @RequestBody CoordinatesAddRequest coordinatesAddRequest
    );

}