# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: IS Project API
  version: 1.0.0
  description: |
    Бэкенд для управления учебными группами и связанными сущностями.
servers:
  - url: /
paths:
  /api/v1/coordinates:
    get:
      tags: [Coordinates]
      summary: Получить список координат
      description: Возвращает постраничный список координат с сортировкой.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortDirectionParam'
      responses:
        '200':
          description: Список координат
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedCoordinatesResponse'
    post:
      tags: [Coordinates]
      summary: Создать координаты
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoordinatesAddRequest'
      responses:
        '200':
          description: Созданные координаты
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoordinatesResponse'
    patch:
      tags: [Coordinates]
      summary: Массово обновить координаты
      parameters:
        - $ref: '#/components/parameters/IdsParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoordinatesUpdateRequest'
      responses:
        '200':
          description: Обновлённые координаты
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CoordinatesResponse'
    delete:
      tags: [Coordinates]
      summary: Массово удалить координаты
      parameters:
        - $ref: '#/components/parameters/IdsParam'
      responses:
        '204':
          description: Координаты удалены
  /api/v1/coordinates/{id}:
    get:
      tags: [Coordinates]
      summary: Получить координаты по идентификатору
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Найденные координаты
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoordinatesResponse'
    patch:
      tags: [Coordinates]
      summary: Обновить координаты
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoordinatesUpdateRequest'
      responses:
        '200':
          description: Обновлённые координаты
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoordinatesResponse'
    delete:
      tags: [Coordinates]
      summary: Удалить координаты
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Удалённые координаты
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoordinatesResponse'
  /api/v1/coordinates/by-ids:
    get:
      tags: [Coordinates]
      summary: Получить несколько координат
      parameters:
        - $ref: '#/components/parameters/IdsParam'
      responses:
        '200':
          description: Список координат
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CoordinatesResponse'
...
components:
  parameters:
    IdPathParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    IdsParam:
      name: ids
      in: query
      required: true
      schema:
        type: array
        items:
          type: integer
          format: int64
      style: form
      explode: true
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
      description: Номер страницы (0-based)
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
      description: Размер страницы
    SortParam:
      name: sort
      in: query
      schema:
        type: string
      description: Нативный параметр Spring для сортировки
    SortByParam:
      name: sortBy
      in: query
      schema:
        type: string
      description: Поле сортировки
    SortDirectionParam:
      name: direction
      in: query
      schema:
        type: string
        enum: [asc, desc]
      description: Направление сортировки
  schemas:
    CoordinatesAddRequest:
      type: object
      required: [x, y]
      properties:
        x:
          type: integer
          format: int64
        y:
          type: number
          format: float
    CoordinatesUpdateRequest:
      type: object
      properties:
        x:
          type: integer
          format: int64
        y:
          type: number
          format: float
    CoordinatesResponse:
      type: object
      required: [id, x, y]
      properties:
        id:
          type: integer
          format: int64
        x:
          type: integer
          format: int64
        y:
          type: number
          format: float
...
    PagedCoordinatesResponse:
      allOf:
        - $ref: '#/components/schemas/PageModel'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/CoordinatesResponse'
...
    PageModel:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...